rviz_common: Remove variadic macro warning check
Remove the check to conditionally disable the
`-Wgnu-variadic-macro-arguments` compiler diagnostic.  As discussed in
issue 402, the `CheckCXXCompilerFlag` CMake check will always pass for a
`-Wno-` compiler flag due to the way gcc suppresses unrecognized command
diagnostics for flags that disable warning messages.

The comments around the check indicate that the flag should be removed
when possible, so this commit removes it.

Signed-off-by: Rich Mattes <richmattes@gmail.com>